eigen error

+++ core/BarElement.h
@@ -21,8 +21,8 @@ using namespace Eigen;
 class BarElement : public FElement {
 
 private:
-    double I_YMZ[4] = { 1, 5, 7, 11 };	// y軸方向,z軸周り成分のインデックス
-    double I_ZMY[4] = { 2, 4, 8, 10 };	// z軸方向,y軸周り成分のインデックス
+    int I_YMZ[4] = { 1, 5, 7, 11 };	// y軸方向,z軸周り成分のインデックス
+    int I_ZMY[4] = { 2, 4, 8, 10 };	// z軸方向,y軸周り成分のインデックス
 
     int param;
     bool isBar;
